tests, net, hotplug: Move 'removeInterface' close to its usage
